refactor: 이름 래퍼 클래스로 변경
