+++ src/main/java/domain/Name.java
@@ -1,0 +1,25 @@
+package domain;
+
+public class Name {
+
+    private static final int NAME_LOWER_BOUND = 0;
+    private static final int NAME_UPPER_BOUND = 5;
+    private static final String INVALID_NAME_LENGTH_MESSAGE = "차의 이름은 1자 이상 ~ 5자 이하여야 합니다.";
+
+    private final String value;
+
+    public Name(final String name) {
+        validate(name);
+        this.value = name;
+    }
+
+    private void validate(String name) {
+        if (NAME_LOWER_BOUND == name.length() || NAME_UPPER_BOUND < name.length()) {
+            throw new IllegalArgumentException(INVALID_NAME_LENGTH_MESSAGE);
+        }
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
