+++ src/test/java/domain/CarTest.java
@@ -1,24 +1,13 @@
 package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 public class CarTest {
-
-    @ParameterizedTest
-    @ValueSource(strings = {"dazzle", ""})
-    @DisplayName("생성자는 이름의 길이가 없거나 5보다 크다면 예외를 던진다.")
-    void should_throwException_when_invalidNameLength(String name) {
-        assertThatThrownBy(() -> new Car(name))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("차의 이름은 1자 이상 ~ 5자 이하여야 합니다.");
-    }
 
     @ParameterizedTest
     @CsvSource(value = {"4,1", "3,0"})
